refactor : mypage파일들 폴더경로 정리
