+++ app/src/main/AndroidManifest.xml
@@ -24,6 +24,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".ui.mypage.ImWriteActivity" />
+
+
     </application>
 
 </manifest>
