+++ app/src/main/java/com/goorm/kkiri/ui/mypage/HelpedListFragment.kt
@@ -1,4 +1,4 @@
-package com.goorm.kkiri.ui.mypage.fragment
+package com.goorm.kkiri.ui.mypage
 import com.goorm.kkiri.R
 import com.goorm.kkiri.base.BaseFragment
 import com.goorm.kkiri.databinding.FragmentHelpedListBinding
