+++ app/src/main/java/com/goorm/kkiri/ui/mypage/ImWriteActivity.kt
@@ -1,9 +1,9 @@
-package com.goorm.kkiri.ui.mypage.Activity
+package com.goorm.kkiri.ui.mypage
 
 import com.goorm.kkiri.R
 import com.goorm.kkiri.base.BaseActivity
 import com.goorm.kkiri.databinding.ActivityImWriteBinding
-import com.goorm.kkiri.ui.mypage.fragment.ImWriteFragment
+
 
 class ImWriteActivity : BaseActivity<ActivityImWriteBinding>(R.layout.activity_im_write) {
