+++ app/src/main/res/layout/activity_im_write.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".ui.mypage.Activity.ImWriteActivity">
+        tools:context=".ui.mypage.fragment.ImWriteActivity">
         <ImageView
             android:id="@+id/imageView2"
             android:layout_width="wrap_content"
